ShopPackagesPO  updated and two obsolete tests moved to Miscellaneous Package
