+++ src/test/java/Miscellaneous/MakeGroupApptTest.java
@@ -1,4 +1,4 @@
-package EME;
+package Miscellaneous;
 
 import java.io.IOException;
